# nextjs-frontend/app/openapi-client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

/**
 * ActionDetectionDatasetSaveRequest
 */
export type ActionDetectionDatasetSaveRequest = {
  /**
   * Action
   * Action label, for example hello, thanks, or iloveyou.
   */
  action: string;
  /**
   * Sequence
   * Sequence of 30 frames, each frame containing 1662 holistic keypoint values.
   */
  sequence: Array<Array<number>>;
};

/**
 * ActionDetectionDatasetSaveResponse
 */
export type ActionDetectionDatasetSaveResponse = {
  /**
   * Success
   */
  success?: boolean;
  result: ActionDetectionDatasetSaveResult;
};

/**
 * ActionDetectionDatasetSaveResult
 */
export type ActionDetectionDatasetSaveResult = {
  /**
   * Action
   */
  action: string;
  /**
   * Sequence Index
   */
  sequence_index: number;
  /**
   * Frame Count
   */
  frame_count: number;
  /**
   * Saved Dir
   */
  saved_dir: string;
};

/**
 * ActionDetectionModelInfoResponse
 */
export type ActionDetectionModelInfoResponse = {
  /**
   * Success
   */
  success?: boolean;
  result: ActionDetectionModelInfoResult;
};

/**
 * ActionDetectionModelInfoResult
 */
export type ActionDetectionModelInfoResult = {
  /**
   * Model Name
   */
  model_name: string;
  /**
   * Labels
   */
  labels: Array<string>;
  /**
   * Sequence Length
   */
  sequence_length: number;
  /**
   * Feature Size
   */
  feature_size: number;
  /**
   * Sample Count
   */
  sample_count?: number | null;
  /**
   * Test Accuracy
   */
  test_accuracy?: number | null;
  /**
   * Source Data Dir
   */
  source_data_dir?: string | null;
};

/**
 * ActionDetectionRequest
 */
export type ActionDetectionRequest = {
  /**
   * Sequence
   * Sequence of 30 frames, each frame containing 1662 holistic keypoint values.
   */
  sequence: Array<Array<number>>;
};

/**
 * ActionDetectionResponse
 */
export type ActionDetectionResponse = {
  /**
   * Success
   */
  success?: boolean;
  result: ActionDetectionResult;
};

/**
 * ActionDetectionResult
 */
export type ActionDetectionResult = {
  /**
   * Label
   */
  label: string;
  /**
   * Confidence
   */
  confidence: number;
  /**
   * Scores
   */
  scores: {
    [key: string]: number;
  };
};

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
  /**
   * Detail
   */
  detail?: Array<ValidationError>;
};

/**
 * TextToSignRequest
 */
export type TextToSignRequest = {
  /**
   * Text
   * Free-form English text to translate into sign tokens.
   */
  text: string;
};

/**
 * TextToSignResponse
 */
export type TextToSignResponse = {
  /**
   * Success
   */
  success?: boolean;
  result: TextToSignResult;
};

/**
 * TextToSignResult
 */
export type TextToSignResult = {
  /**
   * Normalized Text
   */
  normalized_text: string;
  /**
   * Gloss
   */
  gloss: string;
  /**
   * Matched Actions
   */
  matched_actions: Array<string>;
  /**
   * Unsupported Tokens
   */
  unsupported_tokens: Array<string>;
  /**
   * Segments
   */
  segments: Array<TextToSignSegment>;
};

/**
 * TextToSignSegment
 */
export type TextToSignSegment = {
  /**
   * Source
   */
  source: string;
  /**
   * Gloss
   */
  gloss: string;
  /**
   * Supported
   */
  supported: boolean;
  /**
   * Mode
   */
  mode: string;
};

/**
 * ValidationError
 */
export type ValidationError = {
  /**
   * Location
   */
  loc: Array<string | number>;
  /**
   * Message
   */
  msg: string;
  /**
   * Error Type
   */
  type: string;
  /**
   * Input
   */
  input?: unknown;
  /**
   * Context
   */
  ctx?: {
    [key: string]: unknown;
  };
};

export type ActionDetectionSignToTextData = {
  body: ActionDetectionRequest;
  path?: never;
  query?: never;
  url: "/v1/action-detection/sign-to-text";
};

export type ActionDetectionSignToTextErrors = {
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type ActionDetectionSignToTextError =
  ActionDetectionSignToTextErrors[keyof ActionDetectionSignToTextErrors];

export type ActionDetectionSignToTextResponses = {
  /**
   * Successful Response
   */
  200: ActionDetectionResponse;
};

export type ActionDetectionSignToTextResponse =
  ActionDetectionSignToTextResponses[keyof ActionDetectionSignToTextResponses];

export type ActionDetectionModelInfoData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/v1/action-detection/model-info";
};

export type ActionDetectionModelInfoResponses = {
  /**
   * Successful Response
   */
  200: ActionDetectionModelInfoResponse;
};

export type ActionDetectionModelInfoResponse2 =
  ActionDetectionModelInfoResponses[keyof ActionDetectionModelInfoResponses];

export type ActionDetectionTextToSignData = {
  body: TextToSignRequest;
  path?: never;
  query?: never;
  url: "/v1/action-detection/text-to-sign";
};

export type ActionDetectionTextToSignErrors = {
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type ActionDetectionTextToSignError =
  ActionDetectionTextToSignErrors[keyof ActionDetectionTextToSignErrors];

export type ActionDetectionTextToSignResponses = {
  /**
   * Successful Response
   */
  200: TextToSignResponse;
};

export type ActionDetectionTextToSignResponse =
  ActionDetectionTextToSignResponses[keyof ActionDetectionTextToSignResponses];

export type ActionDetectionSaveDatasetSequenceData = {
  body: ActionDetectionDatasetSaveRequest;
  path?: never;
  query?: never;
  url: "/v1/action-detection/dataset";
};

export type ActionDetectionSaveDatasetSequenceErrors = {
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type ActionDetectionSaveDatasetSequenceError =
  ActionDetectionSaveDatasetSequenceErrors[keyof ActionDetectionSaveDatasetSequenceErrors];

export type ActionDetectionSaveDatasetSequenceResponses = {
  /**
   * Successful Response
   */
  200: ActionDetectionDatasetSaveResponse;
};

export type ActionDetectionSaveDatasetSequenceResponse =
  ActionDetectionSaveDatasetSequenceResponses[keyof ActionDetectionSaveDatasetSequenceResponses];

export type HealthHealthCheckData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/health";
};

export type HealthHealthCheckResponses = {
  /**
   * Successful Response
   */
  200: unknown;
};

export type ClientOptions = {
  baseURL: `${string}://openapi.json` | (string & {});
};